Add UI
Three buttons were introduced, along with a light-up-on-hover function.
The scores are automatically displayed above buttons.
The result of each game is appended below match history.
Upon victory, the winner's score lights up (light blue for player;
crimson for computer).

to add:
A way for the newer score to be added on top of the older one
A reset score button that also clears the match history